// src/api/models/changeStoreAdminPasswordRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Partners
 * Allta Partners 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */

export interface ChangeStoreAdminPasswordRequest {
  /** 비밀번호 */
  password: string;
  /** 새 비밀번호 */
  newPassword: string;
  /** 확인용 비밀번호 */
  confirmPassword: string;
}